Add Rudimentary Game Logic Functions
I've added some code to the JavaScript file. One piece is a small
function that serves to make a "choice" for the computer in the
console based implementation of a Rock Paper Scissors game. I've
begun work on several other functions, none of which are noteworthy
as yet beyond their names.